Intel/benchmarks/minigo/implementations/tensorflow/requirements.txt
absl-py
autopep8>=1.3
fire
google.cloud.logging
google.cloud.bigtable
grpcio-tools
keras
numpy>=1.14.0
protobuf
sgf==0.5
six
tqdm>=4.17
pyasn1>=0.4.1
setuptools>=34.0.0

oauth2client==4.1
